+++ src/Modules/MediaEvents/AdminSchema.php
@@ -87,8 +87,7 @@ final class AdminSchema implements AdminSchemaInterface {
 				default_value: false,
 				label: __( 'HTML5 video and audio events', 'duracelltomi-google-tag-manager' ),
 				description: esc_html__( 'Check this option to include a Tag Manager event when a visitor interacts with a native HTML5 <video> or <audio> player embeded on your site. Each event also populates Google Tag Manager\'s built-in Video variables.', 'duracelltomi-google-tag-manager' ),
-				group: 'players',
-				phase: Field::PHASE_EXPERIMENTAL
+				group: 'players'
 			),
 		);
 	}